backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data processing
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
aiofiles==23.2.1

# HTTP client
httpx==0.26.0
aiohttp==3.9.3

# OpenAI
openai==1.12.0

# Validation and serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Rate limiting
slowapi==0.1.9

# CORS
python-dotenv==1.0.1

# Utilities
tenacity==8.2.3
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
